@@ -54,20 +54,22 @@ class FoodlyApp extends StatefulWidget {
 }
 
 class _FoodlyAppState extends State<FoodlyApp> {
-  StreamSubscription<String> _intentDataStreamSubscription;
+  late StreamSubscription<String> _intentDataStreamSubscription;
   Logger _log = new Logger('FoodlyApp');
-  StreamSubscription<LogRecord> _logStream;
-  StreamSubscription<List<Meal>> _privateMealsStream;
-  List<Meal> _privateMealsStreamValue;
-  StreamSubscription<List<Meal>> _publicMealsStream;
-  List<Meal> _publicMealsStreamValue;
+  late StreamSubscription<LogRecord> _logStream;
+  // ignore: cancel_subscriptions
+  StreamSubscription<List<Meal>>? _privateMealsStream;
+  List<Meal>? _privateMealsStreamValue;
+  // ignore: cancel_subscriptions
+  StreamSubscription<List<Meal>>? _publicMealsStream;
+  List<Meal>? _publicMealsStreamValue;
 
   final _appRouter = AppRouter();
 
   @override
   void dispose() {
-    _privateMealsStream.cancel();
-    _publicMealsStream.cancel();
+    _privateMealsStream!.cancel();
+    _publicMealsStream!.cancel();
     _logStream.cancel();
     _intentDataStreamSubscription.cancel();
     super.dispose();
@@ -135,10 +137,10 @@ class _FoodlyAppState extends State<FoodlyApp> {
   Future<void> _loadActivePlan(BuildContext context) async {
     final currentPlan = context.read(planProvider).state;
     if (currentPlan == null) {
-      String planId = await PlanService.getCurrentPlanId();
+      String? planId = await PlanService.getCurrentPlanId();
 
       if (planId != null && planId.isNotEmpty) {
-        Plan newPlan = await PlanService.getPlanById(planId);
+        Plan newPlan = (await PlanService.getPlanById(planId))!;
         context.read(planProvider).state = newPlan;
       }
     }
@@ -147,7 +149,8 @@ class _FoodlyAppState extends State<FoodlyApp> {
   Future<void> _loadActiveUser(BuildContext context) async {
     final firebaseUser = AuthenticationService.currentUser;
     if (firebaseUser != null) {
-      FoodlyUser user = await FoodlyUserService.getUserById(firebaseUser.uid);
+      FoodlyUser user =
+          (await FoodlyUserService.getUserById(firebaseUser.uid))!;
       context.read(userProvider).state = user;
     }
   }
@@ -166,7 +169,7 @@ class _FoodlyAppState extends State<FoodlyApp> {
   void _streamMeals() {
     if (_privateMealsStream == null && _publicMealsStream == null) {
       _privateMealsStream =
-          MealService.streamPlanMeals(context.read(planProvider).state.id)
+          MealService.streamPlanMeals(context.read(planProvider).state!.id!)
               .listen((meals) {
         _privateMealsStreamValue = meals;
         mergeMealsIntoProvider();
@@ -184,8 +187,8 @@ class _FoodlyAppState extends State<FoodlyApp> {
     _log.finer('Call mergeMealsIntoProvider');
 
     var updatedMeals = [
-      ..._privateMealsStreamValue,
-      ..._publicMealsStreamValue
+      ..._privateMealsStreamValue!,
+      ..._publicMealsStreamValue!
     ];
     updatedMeals = [
       ...{...updatedMeals}
@@ -198,7 +201,6 @@ class _FoodlyAppState extends State<FoodlyApp> {
     _intentDataStreamSubscription =
         ReceiveSharingIntent.getTextStream().listen((String value) {
       if (AuthenticationService.currentUser != null &&
-          value != null &&
           value.startsWith(kChefkochShareEndpoint)) {
         context.router
             .push(MealCreateScreenRoute(id: Uri.encodeComponent(value)));
@@ -208,7 +210,7 @@ class _FoodlyAppState extends State<FoodlyApp> {
     });
 
     // For sharing or opening urls/text coming from outside the app while the app is closed
-    ReceiveSharingIntent.getInitialText().then((String value) {
+    ReceiveSharingIntent.getInitialText().then((String? value) {
       if (AuthenticationService.currentUser != null &&
           value != null &&
           value.startsWith(kChefkochShareEndpoint)) {
@@ -227,7 +229,7 @@ class _FoodlyAppState extends State<FoodlyApp> {
       _log.severe('ERR in InAppUpdate.checkForUpdate()', err);
     });
 
-    if (updateInfo?.updateAvailability == UpdateAvailability.updateAvailable) {
+    if (updateInfo.updateAvailability == UpdateAvailability.updateAvailable) {
       InAppUpdate.startFlexibleUpdate().then((_) {
         InAppUpdate.completeFlexibleUpdate().catchError((err) {
           _log.severe('ERR in InAppUpdate.completeFlexibleUpdate()', err);
@@ -254,6 +256,5 @@ class ScrollBehaviorModified extends ScrollBehavior {
       case TargetPlatform.windows:
         return const ClampingScrollPhysics();
     }
-    return const ClampingScrollPhysics();
   }
 }
