@@ -1,33 +1,31 @@
 class PlanMeal {
-  String id;
+  String? id;
   DateTime date;
   String meal;
   MealType type;
-  List<String> upvotes;
-  List<String> downvotes;
+  List<String>? upvotes;
+  List<String>? downvotes;
 
   PlanMeal({
     this.id,
-    this.date,
-    this.meal,
-    this.type,
+    required this.date,
+    required this.meal,
+    required this.type,
     this.upvotes,
     this.downvotes,
   });
 
   Map<String, dynamic> toMap() {
     return {
-      'date': date?.millisecondsSinceEpoch,
+      'date': date.millisecondsSinceEpoch,
       'meal': meal,
-      'type': type?.toString(),
+      'type': type.toString(),
       'upvotes': upvotes,
       'downvotes': downvotes,
     };
   }
 
   factory PlanMeal.fromMap(String id, Map<String, dynamic> map) {
-    if (map == null) return null;
-
     return PlanMeal(
       id: id,
       date: DateTime.fromMillisecondsSinceEpoch(map['date']),
