@@ -1,7 +1,7 @@
 class ShoppingList {
-  String id;
-  String planId;
-  List<String> meals;
+  String? id;
+  String? planId;
+  List<String>? meals;
 
   ShoppingList({
     this.id,
@@ -17,12 +17,10 @@ class ShoppingList {
   }
 
   factory ShoppingList.fromMap(String id, Map<String, dynamic> map) {
-    if (map == null) return null;
-
     return ShoppingList(
       id: id,
       planId: map['planId'],
-      meals: List<String>.from(map['meals'] ?? ''),
+      meals: List<String>.from(map['meals'] ?? '' as Iterable<dynamic>),
     );
   }
 
