@@ -8,9 +8,9 @@ class ClipShadowPath extends StatelessWidget {
   final Widget child;
 
   ClipShadowPath({
-    @required this.shadow,
-    @required this.clipper,
-    @required this.child,
+    required this.shadow,
+    required this.clipper,
+    required this.child,
   });
 
   @override
@@ -29,7 +29,7 @@ class _ClipShadowPainter extends CustomPainter {
   final Shadow shadow;
   final CustomClipper<Path> clipper;
 
-  _ClipShadowPainter({@required this.shadow, @required this.clipper});
+  _ClipShadowPainter({required this.shadow, required this.clipper});
 
   @override
   void paint(Canvas canvas, Size size) {
