@@ -1,4 +1,3 @@
-import 'package:auto_route/auto_route.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import '../../../app_router.gr.dart';
@@ -13,9 +12,9 @@ class PlanDayCard extends StatelessWidget {
   final List<PlanMeal> meals;
 
   const PlanDayCard({
-    Key key,
-    @required this.date,
-    @required this.meals,
+    Key? key,
+    required this.date,
+    required this.meals,
   }) : super(key: key);
 
   @override
@@ -82,7 +81,7 @@ class PlanDayCard extends StatelessWidget {
     return meals.isEmpty || SettingsService.multipleMealsPerTime;
   }
 
-  Widget _buildAddButton(context, {bool isLunch}) {
+  Widget _buildAddButton(context, {required bool isLunch}) {
     return Center(
       child: Padding(
         padding: const EdgeInsets.all(8.0),
