@@ -1,13 +1,14 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:eva_icons_flutter/eva_icons_flutter.dart';
 import 'package:flutter/material.dart';
+// ignore: implementation_imports
 import 'package:flutter_riverpod/src/provider.dart';
 import 'package:flutter_snake_navigationbar/flutter_snake_navigationbar.dart';
-import '../../providers/state_providers.dart';
-import '../../services/shopping_list_service.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 
 import '../../app_router.gr.dart';
+import '../../providers/state_providers.dart';
+import '../../services/shopping_list_service.dart';
 import 'meal_list_view/meal_list_view.dart';
 import 'plan_view/plan_tab_view.dart';
 import 'settings_view/settings_view.dart';
@@ -24,7 +25,7 @@ class _TabNavigationViewState extends State<TabNavigationView> {
   int _currentIndex = 1;
   bool _navbarAnimating = false;
 
-  String _activeListId;
+  String? _activeListId;
 
   @override
   Widget build(BuildContext context) {
@@ -112,7 +113,7 @@ class _TabNavigationViewState extends State<TabNavigationView> {
 
   void _newGrocery() async {
     if (_activeListId == null) {
-      final planId = context.read(planProvider).state.id;
+      final planId = context.read(planProvider).state!.id!;
       final list = await ShoppingListService.getShoppingListByPlanId(planId);
       _activeListId = list.id;
     }
@@ -125,7 +126,7 @@ class _TabNavigationViewState extends State<TabNavigationView> {
       ),
       context: context,
       builder: (modalContext) => EditGroceryModal(
-        shoppingListId: _activeListId,
+        shoppingListId: _activeListId!,
       ),
     );
   }
