@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 
 class SmallCircularProgressIndicator extends StatelessWidget {
-  final Color color;
+  final Color? color;
   final double additionalSize;
 
   final double _iconHeight = 24.0;
 
   SmallCircularProgressIndicator({
-    Key key,
+    Key? key,
     this.color,
     this.additionalSize = 0,
   }) : super(key: key);
@@ -19,8 +19,8 @@ class SmallCircularProgressIndicator extends StatelessWidget {
       width: _iconHeight - 5 + additionalSize,
       child: CircularProgressIndicator(
         strokeWidth: 2.5,
-        valueColor: new AlwaysStoppedAnimation<Color>(
-          color ?? Theme.of(context).textTheme.bodyText1.color,
+        valueColor: new AlwaysStoppedAnimation<Color?>(
+          color ?? Theme.of(context).textTheme.bodyText1!.color,
         ),
       ),
     );
