@@ -52,7 +52,7 @@ class SelectPlanModal extends ConsumerWidget {
               ),
             ),
           ),
-          FutureBuilder<FoodlyUser>(
+          FutureBuilder<FoodlyUser?>(
             future: FoodlyUserService.getUserById(userId),
             builder: (context, snapshot) {
               if (snapshot.connectionState == ConnectionState.waiting) {
@@ -71,7 +71,7 @@ class SelectPlanModal extends ConsumerWidget {
                   ),
                 );
               }
-              return _buildPlanList(snapshot.data, size.height * 0.3);
+              return _buildPlanList(snapshot.data!, size.height * 0.3);
             },
           ),
         ],
@@ -81,7 +81,7 @@ class SelectPlanModal extends ConsumerWidget {
 
   _buildPlanList(FoodlyUser user, double emptySpaceHeight) {
     return FutureBuilder<List<Plan>>(
-      future: PlanService.getPlansByIds(user.oldPlans),
+      future: PlanService.getPlansByIds(user.oldPlans!),
       builder: (context, snapshot) {
         if (snapshot.connectionState == ConnectionState.waiting) {
           return SizedBox(
@@ -102,11 +102,11 @@ class SelectPlanModal extends ConsumerWidget {
 
         return ListView.builder(
           shrinkWrap: true,
-          itemCount: snapshot.data.length,
+          itemCount: snapshot.data!.length,
           itemBuilder: (ctx, index) => ListTile(
-            title: Text(snapshot.data[index].name),
-            subtitle: Text(snapshot.data[index].code),
-            onTap: () => Navigator.pop(ctx, snapshot.data[index]),
+            title: Text(snapshot.data![index].name!),
+            subtitle: Text(snapshot.data![index].code!),
+            onTap: () => Navigator.pop(ctx, snapshot.data![index]),
             trailing: Icon(Icons.arrow_forward_ios_rounded),
           ),
         );
