@@ -11,9 +11,9 @@ class AuthenticationService {
 
   static final log = Logger('AuthenticationService');
   static FirebaseAuth _auth = FirebaseAuth.instance;
-  static User get currentUser => _auth.currentUser;
+  static User? get currentUser => _auth.currentUser;
 
-  static Stream<User> authenticationStream() {
+  static Stream<User?> authenticationStream() {
     log.finer('Stream authStateChanges');
     return _auth.authStateChanges();
   }
@@ -22,7 +22,7 @@ class AuthenticationService {
     log.finer('Call signInWithEmailAndPassword with $email');
     return (await _auth.signInWithEmailAndPassword(
             email: email, password: password))
-        .user
+        .user!
         .uid;
   }
 
@@ -30,7 +30,7 @@ class AuthenticationService {
     log.finer('Call createUserWithEmailAndPassword with $email');
     return (await _auth.createUserWithEmailAndPassword(
             email: email, password: password))
-        .user
+        .user!
         .uid;
   }
 
@@ -87,6 +87,6 @@ class AuthenticationService {
     // not match the nonce in `appleCredential.identityToken`, sign in will fail.
     final user = await _auth.signInWithCredential(oauthCredential);
 
-    return user.user.uid;
+    return user.user!.uid;
   }
 }
