@@ -17,7 +17,7 @@ class ConvertUtil {
   }
 
   /// Converts the amount and the unit (optional) to a good readable string
-  static String amountToString(double amount, [String unit = '']) {
+  static String amountToString(double? amount, [String? unit = '']) {
     String number = amount.toString();
     if (number.endsWith('0')) {
       number = number.substring(0, number.indexOf('.'));
